update shift entry according to changes in value
1. add incl and lesn statement to only look at one file
/etc/shift.conf for speedup
2. add onlyif statement to only execute on changes

+++ code/manifests/shift/entry.pp
@@ -1,12 +1,14 @@
 define lcgdm::shift::entry($component, $type) {
 
     augeas { "shiftentry_${component}-${type}":
-      context => "/files/etc/shift.conf",
+      incl    => "/etc/shift.conf",
+      lens    => "shift.aug",
       changes => [
         "rm name[.='$component'][type='$type']",
         "set name[last()+1] $component",
         "set name[last()]/type $type",
-    ],
+      ],
+      onlyif  => "match name[.='$component'][type='$type'] size == 0",
       require => [ File["/usr/share/augeas/lenses/dist/shift.aug"], File["/etc/shift.conf"], ],
     }
